move es5 dependancies in localtracks.js
