+++ lib/connect.js
@@ -9,7 +9,7 @@ const constants = require('./util/constants');
 const Room = require('./room');
 const E = require('./util/constants').typeErrors;
 const EncodingParametersImpl = require('./encodingparameters');
-const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./index.js');
+const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./localtracks');
 const LocalParticipant = require('./localparticipant');
 const Log = require('./util/log');
 const MediaStreamTrack = require('@twilio/webrtc').MediaStreamTrack;
