+++ lib/createlocaltracks.js
@@ -3,7 +3,7 @@
 const asLocalTrack = require('./util').asLocalTrack;
 const buildLogLevels = require('./util').buildLogLevels;
 const getUserMedia = require('@twilio/webrtc').getUserMedia;
-const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./index.js');
+const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./localtracks');
 const MediaStreamTrack = require('@twilio/webrtc').MediaStreamTrack;
 const Log = require('./util/log');
 const DEFAULT_LOG_LEVEL = require('./util/constants').DEFAULT_LOG_LEVEL;
