+++ lib/index.js
@@ -31,20 +31,19 @@ Object.defineProperties(Video, {
   },
   LocalAudioTrack: {
     enumerable: true,
-    value: require('./media/track/es5/localaudiotrack')
+    value: require('./localtracks').LocalAudioTrack
   },
   LocalDataTrack: {
     enumerable: true,
-    value: require('./media/track/es5/localdatatrack')
+    value: require('./localtracks').LocalDataTrack
   },
   LocalVideoTrack: {
     enumerable: true,
-    value: require('./media/track/es5/localvideotrack')
+    value: require('./localtracks').LocalVideoTrack
   },
   version: {
     enumerable: true,
     value: version
   }
 });
-
 module.exports = Video;
