+++ lib/localparticipant.js
@@ -2,7 +2,7 @@
 
 const util = require('./util');
 const E = require('./util/constants').typeErrors;
-const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./index.js');
+const { LocalAudioTrack, LocalDataTrack, LocalVideoTrack } = require('./localtracks');
 const MediaStreamTrack = require('@twilio/webrtc').MediaStreamTrack;
 const Participant = require('./participant');
 const LocalAudioTrackPublication = require('./media/track/localaudiotrackpublication');
