+++ lib/localtracks.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  LocalAudioTrack: require('./media/track/es5/localaudiotrack'),
+  LocalVideoTrack: require('./media/track/es5/localvideotrack'),
+  LocalDataTrack: require('./media/track/es5/localdatatrack'),
+};
