+++ lib/localtrackses6.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  Audio: require('./media/track/localaudiotrack'),
+  Video: require('./media/track/localvideotrack'),
+  Data: require('./media/track/localdatatrack'),
+};
